test: pin the critique fixes, tighten weak tests
New coverage for every behavior change in the previous commit: the
shutter-close BOTSIM window with unequal exposures, printable-ASCII
refusals at all three sites, the 255-character meta-kernel bound, the
non-finite and malformed offset refusals, the unreadable-instrument
and inverted-range refusals, the all-unreadable exit status, the
ckcls-failure cleanup, and shutter-mode null handling. Review-flagged
weaknesses fixed: thresholds derived from the constants they test,
boundary values pinned just past their bounds, the rounding-tolerance
branch genuinely exercised, hermetic round-trip helper tests moved out
of the holdings guard, kernel-pool isolation extended to the driver
tests, shared helpers moved out of conftest modules, duplicated
fixtures centralized, and the oversized test modules split under the
1000-line cap.
