+++ tests/spindoctor/cli/ck/test_assignment.py
@@ -5,28 +5,39 @@ image really navigated against and one or more holding something else -- and
 lets the assignment step work out which is which by furnishing them and asking.
 The decoys are named so that they would win the tie-break if reproduction were
 skipped, so a test cannot pass by chance of ordering.
+
+The Voyager frozen-attitude lookups have a file of their own,
+``test_assignment_voyager``; what is here is the frame-chain reproduction test,
+the tie-break, the simultaneous-exposure rule, the refusals, and the naming of
+the corrected outputs.
 """
 
-from collections.abc import Callable
 from pathlib import Path
-from typing import Any
 
 import cspyce
 import numpy as np
 import pytest
 from filecache import FCPath
-from tests.spindoctor.cli.ck.conftest import (
+from tests.spindoctor.cli.ck.assignment_helpers import (
+    DECOY_AXIS,
+    DECOY_NAME,
+    IMAGE_NAME,
+    MIDTIME_ET,
+    TRUE_NAME,
+    WRONG_RAD,
+    camera_from_object,
+    cassini_recorded,
+    image_entry,
+    turned,
+    write_candidate,
+)
+from tests.spindoctor.cli.ck.ck_helpers import (
     CASSINI_CAMERA_FRAME,
     CASSINI_CK_FRAME_ID,
     ET0,
-    VOYAGER_CAMERA_FRAME,
-    VOYAGER_CK_FRAME_ID,
     KernelPool,
     axis_rotation,
     baseline_attitude,
-    image_metadata,
-    write_baseline_ck,
-    write_type1_ck,
 )
 
 from spindoctor.cli.ck.assignment import (
@@ -39,181 +50,23 @@ from spindoctor.cli.ck.assignment import (
     reproduces_baseline,
     rotation_angle_rad,
 )
-from spindoctor.cli.ck.images import ImageEntry, OmissionReason
+from spindoctor.cli.ck.images import OmissionReason
 from spindoctor.cli.ck.index import (
-    SNAPPED_LOOKUP_TOL_TICKS,
     CkFile,
     CkIndex,
     CoverageInterval,
     KernelClass,
     build_ck_index,
 )
-from spindoctor.cli.ck.pointing import NDArrayFloatType
 
-# The clocks the two test objects are tagged against, matching the test SCLK.
-_CASSINI_SCLK_ID = -82
-_VOYAGER_SCLK_ID = -31
-
-# Candidate kernels carry half-second records over four seconds from ET0, and
-# the exposure under test sits inside that.
-_RECORD_STEP_S = 0.5
-_RECORDS = 9
-_START_ET = ET0 + 1.0
-_EXPOSURE_S = 2.0
-_MIDTIME_ET = _START_ET + _EXPOSURE_S / 2.0
-
-# Axis for the rotations that turn a baseline into something else.  It is
-# shared with neither the baseline's own axis nor its orientation.
-_DECOY_AXIS = np.array([-0.8, 0.1, 0.59])
-
-# A decoy far outside any tolerance, and the two that bracket the reproduction
-# bound of 1e-9 radians.
-_WRONG_RAD = np.radians(5.0)
+# The two decoy angles that bracket the reproduction bound of 1e-9 radians.
 _JUST_OUTSIDE_RAD = 1e-8
 _WELL_INSIDE_RAD = 1e-10
-
-# Real names from the holdings.  The decoy sorts after the true baseline, so it
-# would win the tie-break if the reproduction test did not exclude it.
-_TRUE_NAME = '03236_04002ra.bc'
-_DECOY_NAME = 'zz04002_04009ra.bc'
-
-_IMAGE_NAME = 'N1484573295_1.IMG'
 
 # An object whose clock id SPICE computes as 0, which no SCLK kernel defines,
 # so a real kernel naming it beside the spacecraft leaves its coverage
 # unreadable.  No image corrects it.
 _CLOCKLESS_OBJECT_ID = -1
-
-
-def _turned(angle_rad: float) -> Callable[[float], NDArrayFloatType]:
-    """Return an attitude history turned from the baseline by a fixed rotation.
-
-    Parameters:
-        angle_rad: How far from the baseline to turn.
-
-    Returns:
-        The attitude at an epoch.
-    """
-    rotation = axis_rotation(_DECOY_AXIS, angle_rad)
-
-    def attitude(et: float) -> NDArrayFloatType:
-        turned: NDArrayFloatType = rotation @ baseline_attitude(et)
-        return turned
-
-    return attitude
-
-
-def _write_candidate(
-    directory: Path,
-    name: str,
-    *,
-    attitude: Callable[[float], NDArrayFloatType] = baseline_attitude,
-    ck_frame_id: int = CASSINI_CK_FRAME_ID,
-    sclk_id: int = _CASSINI_SCLK_ID,
-    start_et: float = ET0,
-    records: int = _RECORDS,
-) -> Path:
-    """Write one candidate C-kernel into a directory, creating the directory.
-
-    Parameters:
-        directory: Directory to write into.
-        name: Basename of the kernel.
-        attitude: The J2000-to-CK-object rotation at an epoch.
-        ck_frame_id: SPICE id of the object the kernel describes.
-        sclk_id: The spacecraft clock its time tags are encoded against.
-        start_et: First record epoch, TDB seconds past J2000.
-        records: Number of half-second records.
-
-    Returns:
-        The path written.
-    """
-    directory.mkdir(parents=True, exist_ok=True)
-    path = directory / name
-    write_baseline_ck(
-        path,
-        ck_frame_id=ck_frame_id,
-        sclk_id=sclk_id,
-        epochs=[start_et + step * _RECORD_STEP_S for step in range(records)],
-        attitude=attitude,
-        angular_velocity=None,
-    )
-    return path
-
-
-def _camera_from_object(ck_frame_id: int, camera_frame: str) -> NDArrayFloatType:
-    """Return the fixed rotation from a CK object's frame to a camera frame.
-
-    Parameters:
-        ck_frame_id: SPICE id of the object.
-        camera_frame: SPICE name of the camera frame.
-
-    Returns:
-        The 3x3 rotation, which the frame kernel defines as a constant.
-    """
-    matrix: NDArrayFloatType = np.asarray(
-        cspyce.pxform(str(cspyce.frmnam(ck_frame_id)), camera_frame, 0.0), dtype=np.float64
-    )
-    return matrix
-
-
-def _entry(
-    *,
-    cmatrix_original: NDArrayFloatType,
-    kernels: tuple[str, ...],
-    image_name: str = _IMAGE_NAME,
-    ck_frame_id: int = CASSINI_CK_FRAME_ID,
-    camera_frame: str = CASSINI_CAMERA_FRAME,
-    start_et: float = _START_ET,
-    **overrides: Any,
-) -> ImageEntry:
-    """Read one image's entry from metadata built around a recorded baseline.
-
-    Parameters:
-        cmatrix_original: The uncorrected attitude the image recorded.
-        kernels: The kernel basenames its provenance recorded.
-        image_name: Basename recorded for the image.
-        ck_frame_id: SPICE id of the object its correction targets.
-        camera_frame: SPICE name of its camera frame.
-        start_et: Exposure start, TDB seconds past J2000.
-        overrides: Further keyword arguments for the metadata builder.
-
-    Returns:
-        The entry.
-    """
-    defaults: dict[str, Any] = {
-        'image_name': image_name,
-        'cmatrix': axis_rotation(_DECOY_AXIS, 0.5) @ np.asarray(cmatrix_original),
-        'cmatrix_original': cmatrix_original,
-        'camera_frame': camera_frame,
-        'ck_frame_id': ck_frame_id,
-        'start_et': start_et,
-        'stop_et': start_et + _EXPOSURE_S,
-        'camera': 'NAC',
-        'kernels': kernels,
-    }
-    defaults.update(overrides)
-    return ImageEntry.from_metadata(image_metadata(**defaults))
-
-
-def _cassini_recorded(pool: KernelPool, *, midtime_et: float = _MIDTIME_ET) -> NDArrayFloatType:
-    """Return the attitude a Cassini image navigated against, in camera terms.
-
-    The frame chain is evaluated the way ``pxform`` evaluates it: the constant
-    rotation from the bus frame to the camera frame, composed onto the bus
-    attitude the baseline kernel holds at the midtime.
-
-    Parameters:
-        pool: The test's kernel pool, which must have the frame kernel
-            furnished.
-        midtime_et: The exposure midtime.
-
-    Returns:
-        The 3x3 J2000-to-camera rotation.
-    """
-    recorded: NDArrayFloatType = _camera_from_object(
-        CASSINI_CK_FRAME_ID, CASSINI_CAMERA_FRAME
-    ) @ baseline_attitude(midtime_et)
-    return recorded
 
 
 def test_the_recorded_baseline_is_what_pxform_gives(pool: KernelPool, tmp_path: Path) -> None:
@@ -222,32 +75,32 @@ def test_the_recorded_baseline_is_what_pxform_gives(pool: KernelPool, tmp_path: 
     Everything below plants ``cmatrix_original`` analytically, so this pins
     that the planted value is what the frame chain actually evaluates to.
     """
-    path = _write_candidate(tmp_path / 'CK-reconstructed', _TRUE_NAME)
+    path = write_candidate(tmp_path / 'CK-reconstructed', TRUE_NAME)
     pool.furnish(path)
     evaluated = np.asarray(
-        cspyce.pxform('J2000', CASSINI_CAMERA_FRAME, _MIDTIME_ET), dtype=np.float64
+        cspyce.pxform('J2000', CASSINI_CAMERA_FRAME, MIDTIME_ET), dtype=np.float64
     )
-    assert rotation_angle_rad(_cassini_recorded(pool), evaluated) < 1e-12
+    assert rotation_angle_rad(cassini_recorded(), evaluated) < 1e-12
 
 
 def test_assignment_keeps_the_reproducing_candidate(pool: KernelPool, tmp_path: Path) -> None:
     """The candidate that answers the recorded attitude is the one assigned."""
     root = tmp_path / 'CK-reconstructed'
-    _write_candidate(root, _TRUE_NAME)
-    _write_candidate(root, _DECOY_NAME, attitude=_turned(_WRONG_RAD))
+    write_candidate(root, TRUE_NAME)
+    write_candidate(root, DECOY_NAME, attitude=turned(WRONG_RAD))
     index = build_ck_index([root])
-    entry = _entry(cmatrix_original=_cassini_recorded(pool), kernels=(_TRUE_NAME, _DECOY_NAME))
+    entry = image_entry(cmatrix_original=cassini_recorded(), kernels=(TRUE_NAME, DECOY_NAME))
     assignments = assign_images([entry], index)
     assert assignments[0].baseline is not None
-    assert assignments[0].baseline.basename == _TRUE_NAME
+    assert assignments[0].baseline.basename == TRUE_NAME
 
 
 def test_assignment_reports_no_reproducing_baseline(pool: KernelPool, tmp_path: Path) -> None:
     """An image whose baseline no candidate reproduces gets no segment."""
     root = tmp_path / 'CK-reconstructed'
-    _write_candidate(root, _DECOY_NAME, attitude=_turned(_WRONG_RAD))
+    write_candidate(root, DECOY_NAME, attitude=turned(WRONG_RAD))
     index = build_ck_index([root])
-    entry = _entry(cmatrix_original=_cassini_recorded(pool), kernels=(_DECOY_NAME,))
+    entry = image_entry(cmatrix_original=cassini_recorded(), kernels=(DECOY_NAME,))
     assignments = assign_images([entry], index)
     assert assignments[0].baseline is None
     assert assignments[0].omission_reason is OmissionReason.NO_REPRODUCING_BASELINE
@@ -258,9 +111,9 @@ def test_a_baseline_that_drifted_past_the_bound_does_not_reproduce(
 ) -> None:
     """A kernel differing by ten nanoradians is a different kernel."""
     root = tmp_path / 'CK-reconstructed'
-    _write_candidate(root, _TRUE_NAME, attitude=_turned(_JUST_OUTSIDE_RAD))
+    write_candidate(root, TRUE_NAME, attitude=turned(_JUST_OUTSIDE_RAD))
     index = build_ck_index([root])
-    entry = _entry(cmatrix_original=_cassini_recorded(pool), kernels=(_TRUE_NAME,))
+    entry = image_entry(cmatrix_original=cassini_recorded(), kernels=(TRUE_NAME,))
     assignments = assign_images([entry], index)
     assert assignments[0].omission_reason is OmissionReason.NO_REPRODUCING_BASELINE
 
@@ -268,9 +121,9 @@ def test_a_baseline_that_drifted_past_the_bound_does_not_reproduce(
 def test_a_baseline_within_the_bound_reproduces(pool: KernelPool, tmp_path: Path) -> None:
     """Floating-point noise between two evaluations is not a difference."""
     root = tmp_path / 'CK-reconstructed'
-    _write_candidate(root, _TRUE_NAME, attitude=_turned(_WELL_INSIDE_RAD))
+    write_candidate(root, TRUE_NAME, attitude=turned(_WELL_INSIDE_RAD))
     index = build_ck_index([root])
-    entry = _entry(cmatrix_original=_cassini_recorded(pool), kernels=(_TRUE_NAME,))
+    entry = image_entry(cmatrix_original=cassini_recorded(), kernels=(TRUE_NAME,))
     assignments = assign_images([entry], index)
     assert assignments[0].omission_reason is None
 
@@ -280,9 +133,9 @@ def test_a_candidate_not_covering_the_midtime_is_not_assigned(
 ) -> None:
     """A kernel whose window ends before the exposure supplied nothing."""
     root = tmp_path / 'CK-reconstructed'
-    _write_candidate(root, _TRUE_NAME, start_et=ET0 + 100.0)
+    write_candidate(root, TRUE_NAME, start_et=ET0 + 100.0)
     index = build_ck_index([root])
-    entry = _entry(cmatrix_original=_cassini_recorded(pool), kernels=(_TRUE_NAME,))
+    entry = image_entry(cmatrix_original=cassini_recorded(), kernels=(TRUE_NAME,))
     assignments = assign_images([entry], index)
     assert assignments[0].omission_reason is OmissionReason.NO_REPRODUCING_BASELINE
 
@@ -292,9 +145,9 @@ def test_an_image_naming_no_indexed_kernel_is_not_assigned(
 ) -> None:
     """A kernel list naming nothing the index holds resolves to no candidate."""
     root = tmp_path / 'CK-reconstructed'
-    _write_candidate(root, _TRUE_NAME)
+    write_candidate(root, TRUE_NAME)
     index = build_ck_index([root])
-    entry = _entry(cmatrix_original=_cassini_recorded(pool), kernels=('naif0012.tls',))
+    entry = image_entry(cmatrix_original=cassini_recorded(), kernels=('naif0012.tls',))
     assignments = assign_images([entry], index)
     assert assignments[0].omission_reason is OmissionReason.NO_REPRODUCING_BASELINE
 
@@ -305,12 +158,12 @@ def test_the_tie_break_prefers_the_reconstructed_kernel(pool: KernelPool, tmp_pa
     gapfill = tmp_path / 'CK-gapfill'
     predicted = tmp_path / 'CK-predicted'
     # The reconstructed name sorts first, so only the class rank can pick it.
-    _write_candidate(reconstructed, '03236_04002ra.bc')
-    _write_candidate(gapfill, 'zz03001_04001pa_gapfill_v01.bc')
-    _write_candidate(predicted, 'zz04009_04051px.bc')
+    write_candidate(reconstructed, '03236_04002ra.bc')
+    write_candidate(gapfill, 'zz03001_04001pa_gapfill_v01.bc')
+    write_candidate(predicted, 'zz04009_04051px.bc')
     index = build_ck_index([reconstructed, gapfill, predicted])
-    entry = _entry(
-        cmatrix_original=_cassini_recorded(pool),
+    entry = image_entry(
+        cmatrix_original=cassini_recorded(),
         kernels=('03236_04002ra.bc', 'zz03001_04001pa_gapfill_v01.bc', 'zz04009_04051px.bc'),
     )
     assignments = assign_images([entry], index)
@@ -323,11 +176,11 @@ def test_the_tie_break_prefers_the_greatest_basename_within_a_class(
 ) -> None:
     """Two reproducing kernels of one class are separated by their names."""
     root = tmp_path / 'CK-reconstructed'
-    _write_candidate(root, '03236_04002ra.bc')
-    _write_candidate(root, '04002_04009ra.bc')
+    write_candidate(root, '03236_04002ra.bc')
+    write_candidate(root, '04002_04009ra.bc')
     index = build_ck_index([root])
-    entry = _entry(
-        cmatrix_original=_cassini_recorded(pool),
+    entry = image_entry(
+        cmatrix_original=cassini_recorded(),
         kernels=('03236_04002ra.bc', '04002_04009ra.bc'),
     )
     assignments = assign_images([entry], index)
@@ -340,20 +193,20 @@ def test_a_simultaneous_pair_yields_one_segment_and_one_loser(
 ) -> None:
     """Both frames reproduce the same baseline; only the narrow angle one is written."""
     root = tmp_path / 'CK-reconstructed'
-    _write_candidate(root, _TRUE_NAME)
+    write_candidate(root, TRUE_NAME)
     index = build_ck_index([root])
-    recorded = _cassini_recorded(pool)
+    recorded = cassini_recorded()
     entries = [
-        _entry(
+        image_entry(
             cmatrix_original=recorded,
-            kernels=(_TRUE_NAME,),
+            kernels=(TRUE_NAME,),
             image_name='N1484573295_1.IMG',
             camera='NAC',
             shutter_mode='BOTSIM',
         ),
-        _entry(
+        image_entry(
             cmatrix_original=recorded,
-            kernels=(_TRUE_NAME,),
+            kernels=(TRUE_NAME,),
             image_name='W1484573295_1.IMG',
             camera='WAC',
             shutter_mode='BOTSIM',
@@ -365,14 +218,50 @@ def test_a_simultaneous_pair_yields_one_segment_and_one_loser(
     assert assignments[1].omission_reason is OmissionReason.BOTSIM_LOSER
 
 
+def test_a_winner_whose_baseline_does_not_reproduce_suppresses_nothing(
+    pool: KernelPool, tmp_path: Path
+) -> None:
+    """A narrow angle frame that writes nothing yields the bus to its partner.
+
+    The narrow angle frame records an attitude no candidate reproduces, so it
+    receives no segment; its wide angle partner's correction then conflicts
+    with nothing and is written rather than omitted as a loser.
+    """
+    root = tmp_path / 'CK-reconstructed'
+    write_candidate(root, TRUE_NAME)
+    index = build_ck_index([root])
+    recorded = cassini_recorded()
+    drifted = axis_rotation(DECOY_AXIS, WRONG_RAD) @ recorded
+    entries = [
+        image_entry(
+            cmatrix_original=drifted,
+            kernels=(TRUE_NAME,),
+            image_name='N1484573295_1.IMG',
+            camera='NAC',
+            shutter_mode='BOTSIM',
+        ),
+        image_entry(
+            cmatrix_original=recorded,
+            kernels=(TRUE_NAME,),
+            image_name='W1484573295_1.IMG',
+            camera='WAC',
+            shutter_mode='BOTSIM',
+        ),
+    ]
+    assignments = assign_images(entries, index)
+    assert assignments[0].omission_reason is OmissionReason.NO_REPRODUCING_BASELINE
+    assert assignments[1].baseline is not None
+    assert assignments[1].omission_reason is None
+
+
 def test_a_fitted_rotation_is_omitted_as_unsupported(pool: KernelPool, tmp_path: Path) -> None:
     """An image with a fitted rotation is reported without being looked up."""
     root = tmp_path / 'CK-reconstructed'
-    _write_candidate(root, _TRUE_NAME)
+    write_candidate(root, TRUE_NAME)
     index = build_ck_index([root])
-    entry = _entry(
-        cmatrix_original=_cassini_recorded(pool),
-        kernels=(_TRUE_NAME,),
+    entry = image_entry(
+        cmatrix_original=cassini_recorded(),
+        kernels=(TRUE_NAME,),
         cmatrix=None,
         rotation_deg=0.25,
     )
@@ -383,11 +272,11 @@ def test_a_fitted_rotation_is_omitted_as_unsupported(pool: KernelPool, tmp_path:
 def test_an_ineligible_image_keeps_its_own_reason(pool: KernelPool, tmp_path: Path) -> None:
     """An image that did not navigate is reported as not eligible."""
     root = tmp_path / 'CK-reconstructed'
-    _write_candidate(root, _TRUE_NAME)
+    write_candidate(root, TRUE_NAME)
     index = build_ck_index([root])
-    entry = _entry(
-        cmatrix_original=_cassini_recorded(pool),
-        kernels=(_TRUE_NAME,),
+    entry = image_entry(
+        cmatrix_original=cassini_recorded(),
+        kernels=(TRUE_NAME,),
         cmatrix=None,
         status='failed',
     )
@@ -400,25 +289,27 @@ def test_assignments_come_back_in_the_order_the_images_were_given(
 ) -> None:
     """Every image considered appears exactly once, in order."""
     root = tmp_path / 'CK-reconstructed'
-    _write_candidate(root, _TRUE_NAME)
+    write_candidate(root, TRUE_NAME)
     index = build_ck_index([root])
-    recorded = _cassini_recorded(pool)
+    recorded = cassini_recorded()
     names = ('N1484573295_1.IMG', 'N1484573296_1.IMG', 'N1484573297_1.IMG')
     entries = [
-        _entry(cmatrix_original=recorded, kernels=(_TRUE_NAME,), image_name=name) for name in names
+        image_entry(cmatrix_original=recorded, kernels=(TRUE_NAME,), image_name=name)
+        for name in names
     ]
     assignments = assign_images(entries, index)
     assert tuple(assignment.image_name for assignment in assignments) == names
 
 
 def test_assign_refuses_two_images_with_the_same_name(pool: KernelPool, tmp_path: Path) -> None:
-    """Two records of one name cannot be told apart in the report."""
+    """Two records of one name cannot be told apart in the report, and are named."""
     root = tmp_path / 'CK-reconstructed'
-    _write_candidate(root, _TRUE_NAME)
+    write_candidate(root, TRUE_NAME)
     index = build_ck_index([root])
-    entry = _entry(cmatrix_original=_cassini_recorded(pool), kernels=(_TRUE_NAME,))
-    with pytest.raises(ValueError, match='two images have the same name'):
+    entry = image_entry(cmatrix_original=cassini_recorded(), kernels=(TRUE_NAME,))
+    with pytest.raises(ValueError, match='named more than once') as refusal:
         assign_images([entry, entry], index)
+    assert IMAGE_NAME in str(refusal.value)
 
 
 def test_assign_refuses_a_pool_that_already_holds_a_kernel(
@@ -426,9 +317,9 @@ def test_assign_refuses_a_pool_that_already_holds_a_kernel(
 ) -> None:
     """A stray C-kernel would answer the lookups alongside the candidate."""
     root = tmp_path / 'CK-reconstructed'
-    path = _write_candidate(root, _TRUE_NAME)
+    path = write_candidate(root, TRUE_NAME)
     index = build_ck_index([root])
-    entry = _entry(cmatrix_original=_cassini_recorded(pool), kernels=(_TRUE_NAME,))
+    entry = image_entry(cmatrix_original=cassini_recorded(), kernels=(TRUE_NAME,))
     pool.furnish(path)
     with pytest.raises(ValueError, match='already furnished'):
         assign_images([entry], index)
@@ -445,7 +336,7 @@ def test_assign_refuses_an_object_whose_coverage_the_index_could_not_read(
     changed since navigation ran.
     """
     root = tmp_path / 'CK-reconstructed'
-    path = _write_candidate(root, _TRUE_NAME)
+    path = write_candidate(root, TRUE_NAME)
     index = CkIndex(
         files=(
             CkFile(
@@ -456,7 +347,7 @@ def test_assign_refuses_an_object_whose_coverage_the_index_could_not_read(
             ),
         )
     )
-    entry = _entry(cmatrix_original=_cassini_recorded(pool), kernels=(_TRUE_NAME,))
+    entry = image_entry(cmatrix_original=cassini_recorded(), kernels=(TRUE_NAME,))
     with pytest.raises(ValueError, match='could not read the coverage'):
         assign_images([entry], index)
 
@@ -466,7 +357,7 @@ def test_assign_ignores_an_unreadable_object_no_image_needs(
 ) -> None:
     """An object nothing corrects does not stop a run, however unreadable it is."""
     root = tmp_path / 'CK-reconstructed'
-    path = _write_candidate(root, _TRUE_NAME)
+    path = write_candidate(root, TRUE_NAME)
     covering = build_ck_index([root]).files[0].coverage
     index = CkIndex(
         files=(
@@ -478,7 +369,7 @@ def test_assign_ignores_an_unreadable_object_no_image_needs(
             ),
         )
     )
-    entry = _entry(cmatrix_original=_cassini_recorded(pool), kernels=(_TRUE_NAME,))
+    entry = image_entry(cmatrix_original=cassini_recorded(), kernels=(TRUE_NAME,))
     assignments = assign_images([entry], index)
     assert assignments[0].baseline is not None
 
@@ -492,7 +383,7 @@ def test_an_image_with_no_pointing_names_no_object_to_read(
     could not read some object's coverage does not turn it into a refusal.
     """
     root = tmp_path / 'CK-reconstructed'
-    path = _write_candidate(root, _TRUE_NAME)
+    path = write_candidate(root, TRUE_NAME)
     index = CkIndex(
         files=(
             CkFile(
@@ -503,9 +394,9 @@ def test_an_image_with_no_pointing_names_no_object_to_read(
             ),
         )
     )
-    entry = _entry(
-        cmatrix_original=_cassini_recorded(pool),
-        kernels=(_TRUE_NAME,),
+    entry = image_entry(
+        cmatrix_original=cassini_recorded(),
+        kernels=(TRUE_NAME,),
         cmatrix=None,
         status='failed',
     )
@@ -516,160 +407,16 @@ def test_an_image_with_no_pointing_names_no_object_to_read(
 def test_the_candidate_pool_is_left_as_it_was_found(pool: KernelPool, tmp_path: Path) -> None:
     """Every candidate furnished for a test is unloaded again."""
     root = tmp_path / 'CK-reconstructed'
-    _write_candidate(root, _TRUE_NAME)
+    write_candidate(root, TRUE_NAME)
     index = build_ck_index([root])
-    entry = _entry(cmatrix_original=_cassini_recorded(pool), kernels=(_TRUE_NAME,))
+    entry = image_entry(cmatrix_original=cassini_recorded(), kernels=(TRUE_NAME,))
     assign_images([entry], index)
     assert int(cspyce.ktotal('CK')) == 0
 
 
-def _voyager_recorded(snapped_et: float) -> NDArrayFloatType:
-    """Return the attitude a Voyager image froze, in camera terms.
-
-    Parameters:
-        snapped_et: The epoch the pointing lookup actually answered at.
-
-    Returns:
-        The 3x3 J2000-to-camera rotation.
-    """
-    recorded: NDArrayFloatType = _camera_from_object(
-        VOYAGER_CK_FRAME_ID, VOYAGER_CAMERA_FRAME
-    ) @ baseline_attitude(snapped_et)
-    return recorded
-
-
-def _snapped_et(midtime_et: float) -> float:
-    """Return the epoch the whole-tick pointing lookup lands on.
-
-    Parameters:
-        midtime_et: The exposure midtime.
-
-    Returns:
-        TDB seconds past J2000 of the encoded whole tick.
-    """
-    return float(cspyce.sct2e(_VOYAGER_SCLK_ID, float(cspyce.sce2t(_VOYAGER_SCLK_ID, midtime_et))))
-
-
-# A midtime that does not land on a whole clock tick, so the snapped lookup
-# answers at a measurably different epoch from the midtime itself.
-_VOYAGER_START_ET = ET0 + 1.003
-_VOYAGER_MIDTIME_ET = _VOYAGER_START_ET + _EXPOSURE_S / 2.0
-
-
-def test_a_frozen_baseline_reproduces_through_the_snapped_lookup(
-    pool: KernelPool, tmp_path: Path
-) -> None:
-    """A Voyager baseline is reproduced at the whole tick, not at the midtime."""
-    root = tmp_path / 'CK'
-    _write_candidate(
-        root,
-        'vg1_sat_version1_type1_iss_sedr.bc',
-        ck_frame_id=VOYAGER_CK_FRAME_ID,
-        sclk_id=_VOYAGER_SCLK_ID,
-    )
-    index = build_ck_index([root])
-    entry = _entry(
-        cmatrix_original=_voyager_recorded(_snapped_et(_VOYAGER_MIDTIME_ET)),
-        kernels=('vg1_sat_version1_type1_iss_sedr.bc',),
-        image_name='C1205021_CALIB.IMG',
-        ck_frame_id=VOYAGER_CK_FRAME_ID,
-        camera_frame=VOYAGER_CAMERA_FRAME,
-        start_et=_VOYAGER_START_ET,
-    )
-    assignments = assign_images([entry], index)
-    assert assignments[0].baseline is not None
-
-
-def test_a_frozen_baseline_read_at_the_midtime_reproduces_through_the_wider_lookup(
-    pool: KernelPool, tmp_path: Path
-) -> None:
-    """The wider lookup asks at the continuous tick, so a midtime reading pairs too.
-
-    The frame oops falls back to encodes the epoch continuously rather than
-    rounding it to a whole tick, and a baseline that interpolates between its
-    records answers such a request at the epoch itself.  An image navigated
-    through that fallback therefore records the attitude at its own midtime,
-    and the second attempt is what pairs it.
-    """
-    root = tmp_path / 'CK'
-    _write_candidate(
-        root,
-        'vg1_sat_version1_type1_iss_sedr.bc',
-        ck_frame_id=VOYAGER_CK_FRAME_ID,
-        sclk_id=_VOYAGER_SCLK_ID,
-    )
-    index = build_ck_index([root])
-    entry = _entry(
-        cmatrix_original=_voyager_recorded(_VOYAGER_MIDTIME_ET),
-        kernels=('vg1_sat_version1_type1_iss_sedr.bc',),
-        image_name='C1205021_CALIB.IMG',
-        ck_frame_id=VOYAGER_CK_FRAME_ID,
-        camera_frame=VOYAGER_CAMERA_FRAME,
-        start_et=_VOYAGER_START_ET,
-    )
-    assignments = assign_images([entry], index)
-    assert assignments[0].baseline is not None
-
-
-def test_a_frozen_baseline_holding_another_attitude_does_not_reproduce(
-    pool: KernelPool, tmp_path: Path
-) -> None:
-    """Neither Voyager lookup pairs a kernel that holds a different attitude."""
-    root = tmp_path / 'CK'
-    _write_candidate(
-        root,
-        'vg1_sat_version1_type1_iss_sedr.bc',
-        ck_frame_id=VOYAGER_CK_FRAME_ID,
-        sclk_id=_VOYAGER_SCLK_ID,
-        attitude=_turned(_WRONG_RAD),
-    )
-    index = build_ck_index([root])
-    entry = _entry(
-        cmatrix_original=_voyager_recorded(_snapped_et(_VOYAGER_MIDTIME_ET)),
-        kernels=('vg1_sat_version1_type1_iss_sedr.bc',),
-        image_name='C1205021_CALIB.IMG',
-        ck_frame_id=VOYAGER_CK_FRAME_ID,
-        camera_frame=VOYAGER_CAMERA_FRAME,
-        start_et=_VOYAGER_START_ET,
-    )
-    assignments = assign_images([entry], index)
-    assert assignments[0].omission_reason is OmissionReason.NO_REPRODUCING_BASELINE
-
-
-def test_a_frozen_baseline_outside_the_first_tolerance_reproduces_at_the_second(
-    pool: KernelPool, tmp_path: Path
-) -> None:
-    """An image navigated through the wider fallback tolerance still pairs.
-
-    The kernel's first record is ten seconds after the exposure, which is
-    beyond the tolerance the primary lookup uses and inside the one the frame
-    oops falls back to; the attitude that lookup answers with is the one at the
-    kernel's first record.
-    """
-    root = tmp_path / 'CK'
-    _write_candidate(
-        root,
-        'vg1_sat_version1_type1_iss_sedr.bc',
-        ck_frame_id=VOYAGER_CK_FRAME_ID,
-        sclk_id=_VOYAGER_SCLK_ID,
-        start_et=ET0 + 10.0,
-    )
-    index = build_ck_index([root])
-    entry = _entry(
-        cmatrix_original=_voyager_recorded(ET0 + 10.0),
-        kernels=('vg1_sat_version1_type1_iss_sedr.bc',),
-        image_name='C1205021_CALIB.IMG',
-        ck_frame_id=VOYAGER_CK_FRAME_ID,
-        camera_frame=VOYAGER_CAMERA_FRAME,
-        start_et=_VOYAGER_START_ET,
-    )
-    assignments = assign_images([entry], index)
-    assert assignments[0].baseline is not None
-
-
 def test_reproduces_baseline_is_false_when_nothing_is_furnished(pool: KernelPool) -> None:
     """A lookup the kernels cannot answer is a candidate that did not supply it."""
-    entry = _entry(cmatrix_original=_cassini_recorded(pool), kernels=(_TRUE_NAME,))
+    entry = image_entry(cmatrix_original=cassini_recorded(), kernels=(TRUE_NAME,))
     assert entry.pointing is not None
     assert reproduces_baseline(entry.pointing) is False
 
@@ -730,10 +477,22 @@ def test_output_basename_refuses_a_name_that_is_not_a_kernel(basename: str) -> N
         output_basename(basename)
 
 
-def test_output_basename_refuses_correcting_a_correction() -> None:
-    """Re-running over an output directory would measure against a corrected baseline."""
+@pytest.mark.parametrize(
+    'basename',
+    ['03236_04002ra_nav.bc', '03236_04002RA_NAV.BC'],
+    ids=['lower-case-marker', 'upper-case-marker'],
+)
+def test_output_basename_refuses_correcting_a_correction(basename: str) -> None:
+    """Re-running over an output directory would measure against a corrected baseline.
+
+    The marker is read case-blind, as the index reads it: an upper-cased copy
+    of a corrected kernel is still a corrected kernel.
+
+    Parameters:
+        basename: A corrected kernel's name, in either case.
+    """
     with pytest.raises(ValueError, match='already a corrected kernel'):
-        output_basename('03236_04002ra_nav.bc')
+        output_basename(basename)
 
 
 def test_an_assignment_names_the_file_its_segment_goes_into(
@@ -741,9 +500,9 @@ def test_an_assignment_names_the_file_its_segment_goes_into(
 ) -> None:
     """An assigned image reports the corrected file that will carry it."""
     root = tmp_path / 'CK-reconstructed'
-    _write_candidate(root, _TRUE_NAME)
+    write_candidate(root, TRUE_NAME)
     index = build_ck_index([root])
-    entry = _entry(cmatrix_original=_cassini_recorded(pool), kernels=(_TRUE_NAME,))
+    entry = image_entry(cmatrix_original=cassini_recorded(), kernels=(TRUE_NAME,))
     assignments = assign_images([entry], index)
     assert assignments[0].output_name == '03236_04002ra_nav.bc'
 
@@ -751,9 +510,9 @@ def test_an_assignment_names_the_file_its_segment_goes_into(
 def test_an_omitted_image_names_no_file(pool: KernelPool, tmp_path: Path) -> None:
     """An image with no baseline goes into no corrected file at all."""
     root = tmp_path / 'CK-reconstructed'
-    _write_candidate(root, _DECOY_NAME, attitude=_turned(_WRONG_RAD))
+    write_candidate(root, DECOY_NAME, attitude=turned(WRONG_RAD))
     index = build_ck_index([root])
-    entry = _entry(cmatrix_original=_cassini_recorded(pool), kernels=(_DECOY_NAME,))
+    entry = image_entry(cmatrix_original=cassini_recorded(), kernels=(DECOY_NAME,))
     assignments = assign_images([entry], index)
     assert assignments[0].output_name is None
 
@@ -763,11 +522,11 @@ def test_group_for_output_carries_every_image_of_one_baseline(
 ) -> None:
     """One corrected file mirrors one original and holds all of its images."""
     root = tmp_path / 'CK-reconstructed'
-    _write_candidate(root, _TRUE_NAME)
+    write_candidate(root, TRUE_NAME)
     index = build_ck_index([root])
-    recorded = _cassini_recorded(pool)
+    recorded = cassini_recorded()
     entries = [
-        _entry(cmatrix_original=recorded, kernels=(_TRUE_NAME,), image_name=name)
+        image_entry(cmatrix_original=recorded, kernels=(TRUE_NAME,), image_name=name)
         for name in ('N1484573295_1.IMG', 'N1484573296_1.IMG')
     ]
     groups = group_for_output(assign_images(entries, index))
@@ -779,9 +538,9 @@ def test_group_for_output_carries_every_image_of_one_baseline(
 def test_group_for_output_skips_omitted_images(pool: KernelPool, tmp_path: Path) -> None:
     """An original no image reproduces yields no file at all."""
     root = tmp_path / 'CK-reconstructed'
-    _write_candidate(root, _DECOY_NAME, attitude=_turned(_WRONG_RAD))
+    write_candidate(root, DECOY_NAME, attitude=turned(WRONG_RAD))
     index = build_ck_index([root])
-    entry = _entry(cmatrix_original=_cassini_recorded(pool), kernels=(_DECOY_NAME,))
+    entry = image_entry(cmatrix_original=cassini_recorded(), kernels=(DECOY_NAME,))
     assert group_for_output(assign_images([entry], index)) == ()
 
 
@@ -791,19 +550,19 @@ def test_group_for_output_refuses_two_originals_with_one_corrected_name(
     """Two directories holding one basename would write one file twice."""
     first = tmp_path / 'CK-reconstructed'
     second = tmp_path / 'CK-gapfill'
-    _write_candidate(first, _TRUE_NAME)
-    _write_candidate(second, _TRUE_NAME, attitude=_turned(_WRONG_RAD))
+    write_candidate(first, TRUE_NAME)
+    write_candidate(second, TRUE_NAME, attitude=turned(WRONG_RAD))
     index = build_ck_index([first, second])
     entries = [
-        _entry(
-            cmatrix_original=_cassini_recorded(pool),
-            kernels=(_TRUE_NAME,),
+        image_entry(
+            cmatrix_original=cassini_recorded(),
+            kernels=(TRUE_NAME,),
             image_name='N1484573295_1.IMG',
         ),
-        _entry(
-            cmatrix_original=_camera_from_object(CASSINI_CK_FRAME_ID, CASSINI_CAMERA_FRAME)
-            @ _turned(_WRONG_RAD)(_MIDTIME_ET),
-            kernels=(_TRUE_NAME,),
+        image_entry(
+            cmatrix_original=camera_from_object(CASSINI_CK_FRAME_ID, CASSINI_CAMERA_FRAME)
+            @ turned(WRONG_RAD)(MIDTIME_ET),
+            kernels=(TRUE_NAME,),
             image_name='N1484573296_1.IMG',
         ),
     ]
@@ -814,21 +573,21 @@ def test_group_for_output_refuses_two_originals_with_one_corrected_name(
 
 def test_an_assignment_needs_a_baseline_or_a_reason(pool: KernelPool) -> None:
     """An assignment that is neither written nor omitted has no disposition."""
-    entry = _entry(cmatrix_original=_cassini_recorded(pool), kernels=(_TRUE_NAME,))
+    entry = image_entry(cmatrix_original=cassini_recorded(), kernels=(TRUE_NAME,))
     with pytest.raises(ValueError, match='not both and not neither'):
         Assignment(entry=entry, baseline=None, omission_reason=None)
 
 
 def test_an_assignment_cannot_write_an_image_with_no_pointing(pool: KernelPool) -> None:
     """A baseline assigned to an ineligible image would have nothing to write."""
-    entry = _entry(
-        cmatrix_original=_cassini_recorded(pool),
-        kernels=(_TRUE_NAME,),
+    entry = image_entry(
+        cmatrix_original=cassini_recorded(),
+        kernels=(TRUE_NAME,),
         cmatrix=None,
         status='failed',
     )
     ck_file = CkFile(
-        path=FCPath('/holdings/CK-reconstructed') / _TRUE_NAME,
+        path=FCPath('/holdings/CK-reconstructed') / TRUE_NAME,
         kernel_class=KernelClass.RECONSTRUCTED,
         coverage=(),
     )
@@ -839,13 +598,13 @@ def test_an_assignment_cannot_write_an_image_with_no_pointing(pool: KernelPool) 
 def test_attitudes_agreeing_within_the_bound_reproduce() -> None:
     """A rotation smaller than the bound is the same attitude."""
     recorded = baseline_attitude(ET0)
-    assert attitudes_reproduce(recorded, axis_rotation(_DECOY_AXIS, 9e-10) @ recorded) is True
+    assert attitudes_reproduce(recorded, axis_rotation(DECOY_AXIS, 9e-10) @ recorded) is True
 
 
 def test_attitudes_differing_beyond_the_bound_do_not_reproduce() -> None:
     """A rotation larger than the bound is a different attitude."""
     recorded = baseline_attitude(ET0)
-    assert attitudes_reproduce(recorded, axis_rotation(_DECOY_AXIS, 1.1e-9) @ recorded) is False
+    assert attitudes_reproduce(recorded, axis_rotation(DECOY_AXIS, 1.1e-9) @ recorded) is False
 
 
 @pytest.mark.parametrize(
@@ -876,189 +635,18 @@ def test_rotation_angle_refuses_a_matrix_of_the_wrong_shape(shape: tuple[int, ..
         rotation_angle_rad(np.ones(shape), np.eye(3))
 
 
-# A discrete baseline holds pointing only at its records, so a lookup answers
-# only within its tolerance of one.  These offsets place the single record
-# where each tolerance can be told apart: inside the base tolerance, and just
-# outside it but inside the term the exposure adds.
-_INSIDE_BASE_TOLERANCE_TICKS = 100.0
-_JUST_OUTSIDE_BASE_TOLERANCE_TICKS = 800.2
-
-# Beyond the base tolerance and beyond the term a long exposure adds to it.
-_BEYOND_THE_EXPOSURE_TERM_TICKS = 801.0
-
-# 800.2 ticks needs 0.2 ticks of exposure term to be reached, which is an
-# exposure of 9.6 s; a Voyager frame's own 0.48 s adds only 0.01.
-_LONG_ENOUGH_EXPOSURE_S = 15.0
-_TOO_SHORT_EXPOSURE_S = 0.48
-
-_VOYAGER_KERNEL_NAME = 'vg1_sat_version1_type1_iss_sedr.bc'
-
-
-def _write_discrete_candidate(
-    directory: Path, name: str, *, offset_ticks: float, midtime_et: float
-) -> Path:
-    """Write a discrete baseline holding one record near an epoch.
-
-    Parameters:
-        directory: Directory to write into.
-        name: Basename of the kernel.
-        offset_ticks: How far after the midtime's whole tick the record sits.
-        midtime_et: The exposure midtime the record is placed relative to.
-
-    Returns:
-        The path written.
-    """
-    directory.mkdir(parents=True, exist_ok=True)
-    path = directory / name
-    tick = float(cspyce.sce2t(_VOYAGER_SCLK_ID, midtime_et)) + offset_ticks
-    write_type1_ck(
-        path,
-        ck_frame_id=VOYAGER_CK_FRAME_ID,
-        ticks=[tick],
-        attitude=baseline_attitude,
-        sclk_id=_VOYAGER_SCLK_ID,
-    )
-    return path
-
-
-def _discrete_entry(offset_ticks: float, *, exposure_s: float = _EXPOSURE_S) -> ImageEntry:
-    """Build the entry for an image whose baseline holds one nearby record.
-
-    Parameters:
-        offset_ticks: How far after the midtime's whole tick the record sits.
-        exposure_s: The recorded exposure, which widens the lookup tolerance.
-
-    Returns:
-        The entry, recording the attitude at the record's own epoch.
-    """
-    midtime = _VOYAGER_MIDTIME_ET
-    tick = float(cspyce.sce2t(_VOYAGER_SCLK_ID, midtime)) + offset_ticks
-    record_et = float(cspyce.sct2e(_VOYAGER_SCLK_ID, tick))
-    start = midtime - exposure_s / 2.0
-    return _entry(
-        cmatrix_original=_voyager_recorded(record_et),
-        kernels=(_VOYAGER_KERNEL_NAME,),
-        image_name='C1205021_CALIB.IMG',
-        ck_frame_id=VOYAGER_CK_FRAME_ID,
-        camera_frame=VOYAGER_CAMERA_FRAME,
-        start_et=start,
-        stop_et=start + exposure_s,
-        midtime_et=midtime,
-        exposure_s=exposure_s,
-    )
-
-
-def test_a_discrete_baseline_answers_only_within_a_tolerance(
-    pool: KernelPool, tmp_path: Path
-) -> None:
-    """A record a hundred ticks away is out of reach with no tolerance at all.
-
-    This is the premise of the two tests below, and the difference between a
-    discrete baseline and every other kernel these tests write: an
-    interpolating baseline answers any epoch inside its window whatever
-    tolerance is asked for, so nothing about a tolerance can be measured on
-    one.
-    """
-    path = _write_discrete_candidate(
-        tmp_path / 'CK',
-        _VOYAGER_KERNEL_NAME,
-        offset_ticks=_INSIDE_BASE_TOLERANCE_TICKS,
-        midtime_et=_VOYAGER_MIDTIME_ET,
-    )
-    pool.furnish(path)
-    tick = float(cspyce.sce2t(_VOYAGER_SCLK_ID, _VOYAGER_MIDTIME_ET))
-    with pytest.raises(OSError, match='CKINSUFFDATA'):
-        cspyce.ckgp(VOYAGER_CK_FRAME_ID, tick, 0.0, 'J2000')
-
-
-def test_a_discrete_baseline_reproduces_within_the_base_tolerance(
-    pool: KernelPool, tmp_path: Path
-) -> None:
-    """A Voyager image is paired with the discrete baseline its lookup reaches."""
-    root = tmp_path / 'CK'
-    _write_discrete_candidate(
-        root,
-        _VOYAGER_KERNEL_NAME,
-        offset_ticks=_INSIDE_BASE_TOLERANCE_TICKS,
-        midtime_et=_VOYAGER_MIDTIME_ET,
-    )
-    index = build_ck_index([root])
-    assignments = assign_images([_discrete_entry(_INSIDE_BASE_TOLERANCE_TICKS)], index)
-    assert assignments[0].baseline is not None
-
-
-def test_the_snapped_lookup_reaches_a_record_within_its_base_tolerance(
-    pool: KernelPool, tmp_path: Path
-) -> None:
-    """The first of the two lookups answers, not just the wider fallback.
-
-    Both lookups find the same record on a discrete baseline, so which of them
-    reached it is visible only in how many attitudes come back.
-    """
-    path = _write_discrete_candidate(
-        tmp_path / 'CK',
-        _VOYAGER_KERNEL_NAME,
-        offset_ticks=_INSIDE_BASE_TOLERANCE_TICKS,
-        midtime_et=_VOYAGER_MIDTIME_ET,
-    )
-    pool.furnish(path)
-    entry = _discrete_entry(_INSIDE_BASE_TOLERANCE_TICKS)
-    assert entry.pointing is not None
-    assert len(baseline_attitudes(entry.pointing)) == 2
-
-
-@pytest.mark.parametrize(
-    ('offset_ticks', 'exposure_s', 'expected_lookups'),
-    [
-        (_JUST_OUTSIDE_BASE_TOLERANCE_TICKS, _LONG_ENOUGH_EXPOSURE_S, 2),
-        (_JUST_OUTSIDE_BASE_TOLERANCE_TICKS, _TOO_SHORT_EXPOSURE_S, 1),
-        (_BEYOND_THE_EXPOSURE_TERM_TICKS, _LONG_ENOUGH_EXPOSURE_S, 1),
-    ],
-    ids=['term-reaches-it', 'term-too-small', 'term-is-not-larger'],
-)
-def test_the_snapped_tolerance_grows_with_the_exposure(
-    pool: KernelPool,
-    tmp_path: Path,
-    offset_ticks: float,
-    exposure_s: float,
-    expected_lookups: int,
-) -> None:
-    """The exposure term is what reaches a record just beyond the base tolerance.
-
-    A 15 s exposure adds 0.3125 ticks of tolerance, which reaches a record
-    800.2 ticks away and not one 801 ticks away; a 0.48 s exposure adds 0.01
-    and reaches neither.  The third case is what pins the size of the term
-    rather than its presence.
-
-    Parameters:
-        offset_ticks: How far past the midtime's whole tick the record sits.
-        exposure_s: The recorded exposure duration.
-        expected_lookups: How many of the two lookups answer.
-    """
-    path = _write_discrete_candidate(
-        tmp_path / 'CK',
-        _VOYAGER_KERNEL_NAME,
-        offset_ticks=offset_ticks,
-        midtime_et=_VOYAGER_MIDTIME_ET,
-    )
-    pool.furnish(path)
-    entry = _discrete_entry(offset_ticks, exposure_s=exposure_s)
-    assert entry.pointing is not None
-    assert len(baseline_attitudes(entry.pointing)) == expected_lookups
-
-
 def test_a_chain_evaluated_object_has_one_lookup(pool: KernelPool, tmp_path: Path) -> None:
     """An object read by evaluating a frame chain is asked once, not twice."""
-    path = _write_candidate(tmp_path / 'CK-reconstructed', _TRUE_NAME)
+    path = write_candidate(tmp_path / 'CK-reconstructed', TRUE_NAME)
     pool.furnish(path)
-    entry = _entry(cmatrix_original=_cassini_recorded(pool), kernels=(_TRUE_NAME,))
+    entry = image_entry(cmatrix_original=cassini_recorded(), kernels=(TRUE_NAME,))
     assert entry.pointing is not None
     assert len(baseline_attitudes(entry.pointing)) == 1
 
 
 def test_baseline_attitudes_is_empty_when_nothing_answers(pool: KernelPool) -> None:
     """A candidate that answers no lookup contributes no attitude to compare."""
-    entry = _entry(cmatrix_original=_cassini_recorded(pool), kernels=(_TRUE_NAME,))
+    entry = image_entry(cmatrix_original=cassini_recorded(), kernels=(TRUE_NAME,))
     assert entry.pointing is not None
     assert baseline_attitudes(entry.pointing) == ()
 
@@ -1070,9 +658,9 @@ def test_assign_refuses_an_undefined_camera_frame(pool: KernelPool, tmp_path: Pa
     is refused once, before any candidate is tried.
     """
     root = tmp_path / 'CK-reconstructed'
-    _write_candidate(root, _TRUE_NAME)
+    write_candidate(root, TRUE_NAME)
     index = build_ck_index([root])
-    entry = _entry(cmatrix_original=_cassini_recorded(pool), kernels=(_TRUE_NAME,))
+    entry = image_entry(cmatrix_original=cassini_recorded(), kernels=(TRUE_NAME,))
     pool.unload(pool.root / 'test.tf')
     with pytest.raises(ValueError, match='is not defined by the furnished kernels'):
         assign_images([entry], index)
@@ -1081,43 +669,16 @@ def test_assign_refuses_an_undefined_camera_frame(pool: KernelPool, tmp_path: Pa
 def test_assign_refuses_an_undefined_ck_object_frame(pool: KernelPool, tmp_path: Path) -> None:
     """The object a corrected kernel targets has to be a frame the pool knows."""
     root = tmp_path / 'CK-reconstructed'
-    _write_candidate(root, _TRUE_NAME)
+    write_candidate(root, TRUE_NAME)
     index = build_ck_index([root])
-    entry = _entry(
-        cmatrix_original=_cassini_recorded(pool),
-        kernels=(_TRUE_NAME,),
+    entry = image_entry(
+        cmatrix_original=cassini_recorded(),
+        kernels=(TRUE_NAME,),
         ck_frame_id=-98000,
         camera_frame=CASSINI_CAMERA_FRAME,
     )
     with pytest.raises(ValueError, match='has no frame name in the furnished kernels'):
         assign_images([entry], index)
-
-
-def test_an_image_at_the_edge_of_the_snapped_lookup_is_still_assigned(
-    pool: KernelPool, tmp_path: Path
-) -> None:
-    """The coverage filter and the lookup reach as far as each other.
-
-    The record sits a tick inside the widest snapped tolerance, which is
-    nearly the furthest a navigated Voyager frame can have been frozen from.
-    The image has to survive the coverage filter and then reproduce; if the two
-    tolerances ever stopped being one value, it would be dropped before any
-    kernel was furnished and reported as having no baseline.  The last tick is
-    left out because the filter measures from the exposure midtime and the
-    lookup from that midtime rounded to a whole tick, so the two agree only to
-    within half a tick of the extreme edge.
-    """
-    offset_ticks = SNAPPED_LOOKUP_TOL_TICKS - 1.0
-    root = tmp_path / 'CK'
-    _write_discrete_candidate(
-        root,
-        _VOYAGER_KERNEL_NAME,
-        offset_ticks=offset_ticks,
-        midtime_et=_VOYAGER_MIDTIME_ET,
-    )
-    index = build_ck_index([root])
-    assignments = assign_images([_discrete_entry(offset_ticks)], index)
-    assert assignments[0].baseline is not None
 
 
 def test_a_candidate_named_by_a_url_is_fetched_before_it_is_furnished(
@@ -1136,7 +697,7 @@ def test_a_candidate_named_by_a_url_is_fetched_before_it_is_furnished(
     than that a URL-spelled name resolves to something SPICE can open.
     """
     root = tmp_path / 'CK-reconstructed'
-    path = _write_candidate(root, _TRUE_NAME)
+    path = write_candidate(root, TRUE_NAME)
     index = CkIndex(
         files=(
             CkFile(
@@ -1150,6 +711,6 @@ def test_a_candidate_named_by_a_url_is_fetched_before_it_is_furnished(
             ),
         )
     )
-    entry = _entry(cmatrix_original=_cassini_recorded(pool), kernels=(_TRUE_NAME,))
+    entry = image_entry(cmatrix_original=cassini_recorded(), kernels=(TRUE_NAME,))
     assignments = assign_images([entry], index)
     assert assignments[0].baseline is not None
