+++ tests/spindoctor/cli/ck/test_segment.py
@@ -16,7 +16,7 @@ from typing import cast
 import cspyce
 import numpy as np
 import pytest
-from tests.spindoctor.cli.ck.conftest import (
+from tests.spindoctor.cli.ck.ck_helpers import (
     CASSINI_CAMERA_FRAME,
     CASSINI_CK_FRAME_ID,
     ET0,
