+++ tests/spindoctor/nav_orchestrator/test_curator.py
@@ -3,6 +3,7 @@
 import dataclasses
 import json
 import math
+from typing import Any, cast
 
 import numpy as np
 import pytest
@@ -21,6 +22,16 @@ from spindoctor.nav_technique.diagnostics import BodyLimbDiagnostics
 from spindoctor.nav_technique.technique_result import NavTechniqueResult
 from spindoctor.support.cmatrix import AttitudeBaseline, PointingSolution
 from spindoctor.support.status_reason import NavStatusReason
+
+
+def _json_round_trip(result: NavResult) -> dict[str, Any]:
+    """Build the metadata dict for a result and round-trip it through real JSON.
+
+    Asserting on the parsed-back dict proves the metadata is genuinely
+    serializable and pins what a reader of the written file sees, not what the
+    in-memory objects happened to be.
+    """
+    return cast(dict[str, Any], json.loads(json.dumps(build_metadata_dict(result))))
 
 
 def _classifier() -> NavImageClassifierResult:
@@ -271,34 +282,34 @@ def _gated_titan_result() -> NavResult:
 
 def test_gated_feature_entry_reaches_the_json() -> None:
     """A gated feature is recorded in the emitted metadata, not only in the log."""
-    md = json.loads(json.dumps(build_metadata_dict(_gated_titan_result())))
+    md = _json_round_trip(_gated_titan_result())
     entry = md['feature_inventory'][0]
     assert entry['gated'] is True
 
 
 def test_gated_feature_entry_names_its_type() -> None:
     """The gate record identifies which feature type was dropped."""
-    md = json.loads(json.dumps(build_metadata_dict(_gated_titan_result())))
+    md = _json_round_trip(_gated_titan_result())
     assert md['feature_inventory'][0]['feature_type'] == 'TITAN_LIMB'
 
 
 def test_gated_feature_entry_carries_its_breakdown() -> None:
     """The reliability breakdown travels into the JSON so gates are attributable."""
-    md = json.loads(json.dumps(build_metadata_dict(_gated_titan_result())))
+    md = _json_round_trip(_gated_titan_result())
     reasons = md['feature_inventory'][0]['reliability_reasons']
     assert reasons['titan_occluded_fraction'] == pytest.approx(0.42)
 
 
 def test_breakdown_omits_inapplicable_components() -> None:
     """Components that do not apply to a feature type are left out entirely."""
-    md = json.loads(json.dumps(build_metadata_dict(_gated_titan_result())))
+    md = _json_round_trip(_gated_titan_result())
     reasons = md['feature_inventory'][0]['reliability_reasons']
     assert 'predicted_snr' not in reasons
 
 
 def test_breakdown_is_empty_when_no_component_was_populated() -> None:
     """A feature whose model populated no component reports an empty mapping."""
-    md = json.loads(json.dumps(build_metadata_dict(_ok_result_with_one_technique())))
+    md = _json_round_trip(_ok_result_with_one_technique())
     assert md['feature_inventory'][0]['reliability_reasons'] == {}
 
 
@@ -338,6 +349,12 @@ def _pointing(*, corrected: bool) -> PointingSolution:
 
 
 def _result_with_pointing(*, corrected: bool) -> NavResult:
+    """Return the one-technique success result stamped with a pointing solution.
+
+    Parameters:
+        corrected: True to carry a corrected C-matrix alongside the baseline;
+            False for a solution recording only the uncorrected attitude.
+    """
     return dataclasses.replace(
         _ok_result_with_one_technique(), pointing=_pointing(corrected=corrected)
     )
@@ -357,7 +374,7 @@ def test_metadata_dict_omits_times_when_none_was_computed() -> None:
 
 def test_pointing_block_has_the_declared_key_set() -> None:
     """A corrected result's pointing block carries exactly the declared keys."""
-    md = json.loads(json.dumps(build_metadata_dict(_result_with_pointing(corrected=True))))
+    md = _json_round_trip(_result_with_pointing(corrected=True))
     assert set(md['pointing']) == {
         'cmatrix',
         'cmatrix_original',
@@ -369,7 +386,7 @@ def test_pointing_block_has_the_declared_key_set() -> None:
 
 def test_times_block_has_the_declared_key_set() -> None:
     """The times block carries exactly the declared keys."""
-    md = json.loads(json.dumps(build_metadata_dict(_result_with_pointing(corrected=True))))
+    md = _json_round_trip(_result_with_pointing(corrected=True))
     assert set(md['times']) == {
         'start_et',
         'stop_et',
@@ -383,19 +400,19 @@ def test_times_block_has_the_declared_key_set() -> None:
 
 def test_cmatrix_is_serialized_as_nine_row_major_floats() -> None:
     """The corrected C-matrix flattens row by row, not column by column."""
-    md = json.loads(json.dumps(build_metadata_dict(_result_with_pointing(corrected=True))))
+    md = _json_round_trip(_result_with_pointing(corrected=True))
     assert md['pointing']['cmatrix'] == [0.0, 0.0, 1.0, 1.0, 0.0, 0.0, 0.0, 1.0, 0.0]
 
 
 def test_cmatrix_original_is_serialized_as_nine_row_major_floats() -> None:
     """The uncorrected C-matrix flattens row by row too."""
-    md = json.loads(json.dumps(build_metadata_dict(_result_with_pointing(corrected=True))))
+    md = _json_round_trip(_result_with_pointing(corrected=True))
     assert md['pointing']['cmatrix_original'] == [0.0, 1.0, 0.0, 0.0, 0.0, 1.0, 1.0, 0.0, 0.0]
 
 
 def test_frame_identities_are_serialized_as_recorded() -> None:
     """The camera frame name and both frame ids travel unchanged."""
-    md = json.loads(json.dumps(build_metadata_dict(_result_with_pointing(corrected=True))))
+    md = _json_round_trip(_result_with_pointing(corrected=True))
     assert md['pointing']['camera_frame'] == 'CASSINI_ISS_NAC'
     assert md['pointing']['camera_frame_id'] == -82360
     assert md['pointing']['ck_frame_id'] == -82000
@@ -403,7 +420,7 @@ def test_frame_identities_are_serialized_as_recorded() -> None:
 
 def test_times_are_serialized_unrounded() -> None:
     """Epochs keep full precision, since they define a segment interval exactly."""
-    md = json.loads(json.dumps(build_metadata_dict(_result_with_pointing(corrected=True))))
+    md = _json_round_trip(_result_with_pointing(corrected=True))
     assert md['times']['start_et'] == 246684087.05644953
     assert md['times']['midtime_et'] == 246684087.14644954
     assert md['times']['stop_et'] == 246684087.23644954
@@ -411,7 +428,7 @@ def test_times_are_serialized_unrounded() -> None:
 
 def test_sclk_strings_are_serialized_as_recorded() -> None:
     """The three spacecraft-clock strings travel unchanged."""
-    md = json.loads(json.dumps(build_metadata_dict(_result_with_pointing(corrected=True))))
+    md = _json_round_trip(_result_with_pointing(corrected=True))
     assert md['times']['sclk_start'] == '1/1572105349.077'
     assert md['times']['sclk_midtime'] == '1/1572105349.100'
     assert md['times']['sclk_stop'] == '1/1572105349.123'
@@ -419,12 +436,12 @@ def test_sclk_strings_are_serialized_as_recorded() -> None:
 
 def test_uncorrectable_result_omits_cmatrix_but_keeps_the_original() -> None:
     """A solution with no corrected attitude still records the uncorrected one."""
-    md = json.loads(json.dumps(build_metadata_dict(_result_with_pointing(corrected=False))))
+    md = _json_round_trip(_result_with_pointing(corrected=False))
     assert 'cmatrix' not in md['pointing']
     assert md['pointing']['cmatrix_original'] == [0.0, 1.0, 0.0, 0.0, 0.0, 1.0, 1.0, 0.0, 0.0]
 
 
 def test_uncorrectable_result_still_records_its_times() -> None:
     """A solution with no corrected attitude still records the exposure times."""
-    md = json.loads(json.dumps(build_metadata_dict(_result_with_pointing(corrected=False))))
+    md = _json_round_trip(_result_with_pointing(corrected=False))
     assert md['times']['exposure_s'] == 0.18
